changes done for idt, not working yet

+++ include/sys/idt.h
@@ -5,7 +5,7 @@
 
 //bool are_interrupts_enabled();
 //void lidt(void* base, uint16_t size);
-void get_entry_idt(int i,uint64_t target,uint8_t flags);
+void get_entry_idt(uint32_t i,uint64_t target,uint16_t flags);
 void reload_idt();
